Added half star capability for display ratings

+++ BeanSpotted/CoffeeShop.swift
@@ -22,7 +22,7 @@ class CoffeeShop {
     var local: Bool
     private(set) var createTime = Date.now
     var modifyTime = Date.now
-    var avgRating: Int = 0
+    var avgRating: Double = 0
     @Relationship(deleteRule: .cascade, inverse: \Review.coffeeShop) var reviews = [Review]()
     
     init(name: String = "Mom n' Em", address: String = "4310 Whetsel Ave, Cincinnati, OH 45227", openingTime: Date = Date.now, closingTime: Date = Date.now, decafAvailable: Bool = true, local: Bool = true) {
@@ -36,7 +36,7 @@ class CoffeeShop {
     }
     
     // Returns average overall rating of all coffee shop reviews
-    func averageOverallRating() -> Int {
+    func averageOverallRating() -> Double {
         
         // Map reviews array to overall review ratings array
         let ratings = reviews.map( {$0.overallRating} )
@@ -44,7 +44,7 @@ class CoffeeShop {
         if reviews.count > 0 {
             //self.avgRating = ratings.reduce(0, +) / reviews.count
             //return self.avgRating
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
@@ -52,13 +52,13 @@ class CoffeeShop {
     }
     
     // Returns average coffee rating of all coffee shop reviews
-    func averageCoffeeRating() -> Int {
+    func averageCoffeeRating() -> Double {
         
         // Map reviews array to coffee ratings array
         let ratings = reviews.map( {$0.coffee} )
         
         if reviews.count > 0 {
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
@@ -66,13 +66,13 @@ class CoffeeShop {
     }
     
     // Returns average non-coffee drink rating of all coffee shop reviews
-    func averageNonCoffeeDrinkRating() -> Int {
+    func averageNonCoffeeDrinkRating() -> Double {
         
         // Map reviews array to non-coffee drink ratings array
         let ratings = reviews.map( {$0.nonCoffeeDrinks} )
         
         if reviews.count > 0 {
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
@@ -80,13 +80,13 @@ class CoffeeShop {
     }
     
     // Returns average safety rating of all coffee shop reviews
-    func averageSafetyRating() -> Int {
+    func averageSafetyRating() -> Double {
         
         // Map reviews array to safety ratings array
         let ratings = reviews.map( {$0.safety} )
         
         if reviews.count > 0 {
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
@@ -94,13 +94,13 @@ class CoffeeShop {
     }
     
     // Returns average wifi quality rating of all coffee shop reviews
-    func averageWifiQualityRating() -> Int {
+    func averageWifiQualityRating() -> Double {
         
         // Map reviews array to wifi quality ratings array
         let ratings = reviews.map( {$0.wifiQuality} )
         
         if reviews.count > 0 {
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
@@ -108,13 +108,13 @@ class CoffeeShop {
     }
     
     // Returns average seating rating of all coffee shop reviews
-    func averageSeatingRating() -> Int {
+    func averageSeatingRating() -> Double {
         
         // Map reviews array to seating ratings array
         let ratings = reviews.map( {$0.seating} )
         
         if reviews.count > 0 {
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
@@ -122,13 +122,13 @@ class CoffeeShop {
     }
     
     // Returns average quiet rating of all coffee shop reviews
-    func averageQuietRating() -> Int {
+    func averageQuietRating() -> Double {
         
         // Map reviews array to quiet ratings array
         let ratings = reviews.map( {$0.quiet} )
         
         if reviews.count > 0 {
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
@@ -136,13 +136,13 @@ class CoffeeShop {
     }
     
     // Returns average parking rating of all coffee shop reviews
-    func averageParkingRating() -> Int {
+    func averageParkingRating() -> Double {
         
         // Map reviews array to parking ratings array
         let ratings = reviews.map( {$0.parking} )
         
         if reviews.count > 0 {
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
@@ -150,13 +150,13 @@ class CoffeeShop {
     }
     
     // Returns average food rating of all coffee shop reviews
-    func averageFoodRating() -> Int {
+    func averageFoodRating() -> Double {
         
         // Map reviews array to food ratings array
         let ratings = reviews.map( {$0.food} )
         
         if reviews.count > 0 {
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
@@ -164,13 +164,13 @@ class CoffeeShop {
     }
     
     // Returns average value rating of all coffee shop reviews
-    func averageValueRating() -> Int {
+    func averageValueRating() -> Double {
         
         // Map reviews array to value ratings array
         let ratings = reviews.map( {$0.value} )
         
         if reviews.count > 0 {
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
@@ -178,13 +178,13 @@ class CoffeeShop {
     }
     
     // Returns average cleanliness rating of all coffee shop reviews
-    func averageCleanlinessRating() -> Int {
+    func averageCleanlinessRating() -> Double {
         
         // Map reviews array to cleanliness ratings array
         let ratings = reviews.map( {$0.cleanliness} )
         
         if reviews.count > 0 {
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
@@ -192,13 +192,13 @@ class CoffeeShop {
     }
     
     // Returns average staff friendliness rating of all coffee shop reviews
-    func averageStaffFriendlinessRating() -> Int {
+    func averageStaffFriendlinessRating() -> Double {
         
         // Map reviews array to staff friendliness ratings array
         let ratings = reviews.map( {$0.staffFriendliness} )
         
         if reviews.count > 0 {
-            return ratings.reduce(0, +) / reviews.count
+            return ratings.reduce(0, +) / Double(reviews.count)
         } else {
             print("No reviews for this coffee shop name + address")
             return 0
